Add support for pincodes.
This duplicates the functionality of RSA tokens. The basic
idea is that pincodes are stored in the /etc/shadow -style file
with:

user:hashcode:whateverelse

This way existing tools written to work on shadow files can manage this
file.

There are quite a few changes around the code to accommodate this, but
it's not like this is used anywhere. ;)