Exclude build & dist directories & MANIFEST file
The build and dist directories created by setup.py should be ignored as
they and their contents shouldn't ever be checked in.

Additionally MANIFEST is created during the process from MANIFEST.in or
if that is missing from the files utilized and described in setup.py so
we should also ignore it.