Make sure that window_size is defined
It is possible to generate totp files with no WINDOW_SIZE specified.
According to the documenation with Google Authenticator the default
should be 3 if it is not defined. When generating the totp file using
google-authenticator answering yes for increasing the time-skew window
increases it to 4 minutes (WINDOW_SIZE 17).

This is just a safety catch in case the WINDOW_SIZE is not defined.