Update files for package building & bump version
With the change to the backends architecture we need to make sure that
the backends submodule gets pulled into the package build.

As there is a bit of architecture change plus new functionality I'm
bumping the version number.

We also fix up the spec file to deal with the upgrade scenario properly.
Formerly we would delete the selinux policies during upgrade which
should only happen on full uninstall.