Update FreeRADIUS handler to be more generic
FreeRADIUS perl handler now is multi-name entrant in support of symlink
name changes for different config files.

Documentation update on changes dealing with optional group filtering as
well as the expectation that an account that passes the user filter is
considered valid, unless it doesn't pass the optional group filter.