Better guard against replay attacks
https://github.com/mricon/totp-cgi/issues/12

When pincodes are used in the pincode+tokencode configuration, making a
typo in the pincode will not mark the tokencode as used, so an attacker
can reuse the same tokencode with the correct pincode.

This change will always record failed attempts and check against
failed timestamps in addition to successful timestamps, when checking
for token reuse. This basically makes it impossible to reuse a token
even if there was a pincode or token typo in the first instance, or even
if someone typed the absolutely wrong token code.

The downside is that the client will have to always wait until token has
changed before they can proceed, but the upside is that attackers cannot
try to replay failed attempts in any typo scenario.